+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 -   Fix issue with installing Fabric not downloading all libraries needed
 -   Add ability to select loader versions for packs that enable it
 -   Add ability to add mods from within the launcher direct from Curse
+-   Add in selectbox in Java settings tab to Java paths on windows